+++ plone/app/querystring/querybuilder.py
@@ -1,21 +1,23 @@
 import json
+import logging
 
-from zope.component import getMultiAdapter, getUtility
+from zope.component import getMultiAdapter, getUtility, getUtilitiesFor
 from zope.i18n import translate
 
 from zope.i18nmessageid import MessageFactory
 from zope.publisher.browser import BrowserView
 
-
 from plone.app.contentlisting.interfaces import IContentListing
 from plone.registry.interfaces import IRegistry
 from plone.app.querystring import queryparser
+from plone.app.querystring.interfaces import IParsedQueryIndexModifier
 
 from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.PloneBatch import Batch
+from plone.batching import Batch
 
 from .interfaces import IQuerystringRegistryReader
 
+logger = logging.getLogger('plone.app.querystring')
 _ = MessageFactory('plone')
 
 
@@ -36,13 +38,23 @@ class QueryBuilder(BrowserView):
 
     def __call__(self, query, batch=False, b_start=0, b_size=30,
                  sort_on=None, sort_order=None, limit=0, brains=False,
-                 fieldname='', in_factory=False, acquire=False):
+                 fieldname='', in_factory=False, acquire=False,
+                 custom_query={}):
         """If there are results, make the query and return the results"""
         if self._results is None:
-            self._results = self._makequery(query=query, batch=batch,
-                b_start=b_start, b_size=b_size, sort_on=sort_on,
-                sort_order=sort_order, limit=limit, brains=brains,
-                fieldname=fieldname, in_factory=in_factory, acquire=acquire)
+            self._results = self._makequery(
+                query=query,
+                batch=batch,
+                b_start=b_start,
+                b_size=b_size,
+                sort_on=sort_on,
+                sort_order=sort_order,
+                limit=limit,
+                brains=brains,
+                fieldname=fieldname,
+                in_factory=in_factory,
+                acquire=acquire,
+                custom_query=custom_query)
         return self._results
 
     def html_results(self, query, fieldname='', in_factory=False, acquire=False):
@@ -53,23 +65,49 @@ class QueryBuilder(BrowserView):
                        sort_order=self.request.get('sort_order', None),
                        limit=10, fieldname=fieldname, in_factory=in_factory, acquire=acquire)
 
-        return getMultiAdapter((results, self.request),
-            name='display_query_results')(**options)
+        return getMultiAdapter(
+            (results, self.request),
+            name='display_query_results'
+        )(**options)
 
     def _makequery(self, query=None, batch=False, b_start=0, b_size=30,
                    sort_on=None, sort_order=None, limit=0, brains=False,
-                   fieldname='', in_factory=False, acquire=False):
+                   fieldname='', in_factory=False, acquire=False,
+                   custom_query={}):
         """Parse the (form)query and return using multi-adapter"""
         parsedquery = queryparser.parseFormquery(
             self.context, query, sort_on, sort_order, fieldname=fieldname,
             in_factory=in_factory, acquire=acquire)
+
+        index_modifiers = getUtilitiesFor(IParsedQueryIndexModifier)
+        for name, modifier in index_modifiers:
+            if name in parsedquery:
+                new_name, query = modifier(parsedquery[name])
+                parsedquery[name] = query
+                # if a new index name has been returned, we need to replace
+                # the native ones
+                if name != new_name:
+                    del parsedquery[name]
+                    parsedquery[new_name] = query
+
+        # Check for valid indexes
+        catalog = getToolByName(self.context, 'portal_catalog')
+        valid_indexes = [index for index in parsedquery
+                         if index in catalog.indexes()]
+
+        # We'll ignore any invalid index, but will return an empty set if none
+        # of the indexes are valid.
+        if not valid_indexes:
+            logger.warning(
+                "Using empty query because there are no valid indexes used.")
+            parsedquery = {}
+
         if not parsedquery:
             if brains:
                 return []
             else:
                 return IContentListing([])
 
-        catalog = getToolByName(self.context, 'portal_catalog')
         if batch:
             parsedquery['b_start'] = b_start
             parsedquery['b_size'] = b_size
@@ -79,48 +117,32 @@ class QueryBuilder(BrowserView):
         if 'path' not in parsedquery:
             parsedquery['path'] = {'query': ''}
 
-        # The Subject field in Plone currently uses a utf-8 encoded string.
-        # When a catalog query tries to compare a unicode string from the
-        # parsedquery with existing utf-8 encoded string indexes unindexing
-        # will fail with a UnicodeDecodeError. To prevent this from happening
-        # we always encode the Subject query.
-        # XXX: As soon as Plone uses unicode for all indexes, this code can
-        # be removed.
-        if 'Subject' in parsedquery:
-            query = parsedquery['Subject']['query']
-            # query can be a unicode string or a list of unicode strings.
-            if isinstance(query, unicode):
-                parsedquery['Subject']['query'] = query.encode("utf-8")
-            elif isinstance(query, list):
-                # We do not want to change the collections' own query string,
-                # therefore we create a new copy of the list.
-                copy_of_query = list(query)
-                # Iterate over all query items and encode them if they are
-                # unicode strings
-                i = 0
-                for item in copy_of_query:
-                    if isinstance(item, unicode):
-                        copy_of_query[i] = item.encode("utf-8")
-                    i += 1
-                parsedquery['Subject']['query'] = copy_of_query
-            else:
-                pass
+        if isinstance(custom_query, dict):
+            # Update the parsed query with extra query dictionary. This may
+            # override parsed query options.
+            parsedquery.update(custom_query)
 
-        results = catalog(parsedquery)
+        results = catalog(**parsedquery)
+        if getattr(results, 'actual_result_count', False) and limit\
+                and results.actual_result_count > limit:
+            results.actual_result_count = limit
+
         if not brains:
             results = IContentListing(results)
         if batch:
-            results = Batch(results, b_size, b_start)
+            results = Batch(results, b_size, start=b_start)
         return results
 
     def number_of_results(self, query, fieldname='', in_factory=False, acquire=False):
         """Get the number of results"""
         results = self(query, sort_on=None, sort_order=None, limit=1,
                        fieldname=fieldname, in_factory=in_factory, acquire=acquire)
-        return translate(_(u"batch_x_items_matching_your_criteria",
-                 default=u"${number} items matching your search terms.",
-                 mapping={'number': results.actual_result_count}),
-                 context=self.request)
+        return translate(
+            _(u"batch_x_items_matching_your_criteria",
+              default=u"${number} items matching your search terms.",
+              mapping={'number': results.actual_result_count}),
+            context=self.request
+        )
 
 
 class RegistryConfiguration(BrowserView):