+++ plone/app/querystring/tests/testQueryBuilder.py
@@ -1,15 +1,20 @@
  # -*- coding: utf-8 -*-
 
+from plone.app.querystring.testing import TEST_PROFILE_PLONEAPPQUERYSTRING_INTEGRATION_TESTING
+
 from zope.component import getMultiAdapter
 from zope.publisher.browser import TestRequest
 
-from .base import QuerystringTestCase
+import unittest2 as unittest
 
 
-class TestQuerybuilder(QuerystringTestCase):
+class TestQuerybuilder(unittest.TestCase):
 
-    def afterSetUp(self):
-        self.loginAsPortalOwner()
+    layer = TEST_PROFILE_PLONEAPPQUERYSTRING_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+
         self.portal.invokeFactory("Document",
                                   "collectionstestpage",
                                   title="Collectionstestpage")
@@ -17,13 +22,19 @@ class TestQuerybuilder(QuerystringTestCase):
         self.testpage = testpage
         self.portal.portal_workflow.doActionFor(testpage, 'publish')
         self.request = TestRequest()
-        self.querybuilder = getMultiAdapter((self.portal, self.request),
-                                             name='querybuilderresults')
+        self.querybuilder = getMultiAdapter(
+            (self.portal, self.request),
+            name='querybuilderresults'
+        )
         self.query = [{
             'i': 'Title',
             'o': 'plone.app.querystring.operation.string.is',
             'v': 'Collectionstestpage',
         }]
+        self.portal.invokeFactory("Folder",
+                                  "testfolder",
+                                  title="Test Folder")
+        self.folder = self.portal.testfolder
 
     def testQueryBuilderQuery(self):
         results = self.querybuilder(query=self.query)
@@ -136,8 +147,57 @@ class TestQuerybuilder(QuerystringTestCase):
             results[0].getObject().Subject(),
             ('Äüö',))
 
+    def testQueryBuilderCustomQuery(self):
+        """Test, if custom queries are respected when getting the results.
+        """
 
-class TestConfigurationFetcher(QuerystringTestCase):
+        # It would be slightly nicer to compare directly against the changed
+        # query. But instead we have to test for changed results, as _makequery
+        # returns the results but not the query.
+
+        query = [{
+            'i': 'Title',
+            'o': 'plone.app.querystring.operation.string.is',
+            'v': 'Collectionstestpage',
+        }]
+
+        # Test normal, without custom_query.
+        results = self.querybuilder._makequery(query=query)
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].Title(), 'Collectionstestpage')
+
+        # Test with changed ``Title`` query, overwriting original query.
+        results = self.querybuilder._makequery(
+            query=query,
+            custom_query={'Title': {'query': 'Test Folder'}})
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].Title(), 'Test Folder')
+
+        # Test with changed ``portal_type``, but other ``Title`` query.
+        results = self.querybuilder._makequery(
+            query=query,
+            custom_query={'portal_type': {'query': 'Folder'}})
+        self.assertEqual(len(results), 0)
+
+        # Test with changed ``portal_type`` and changed ``Title``.
+        results = self.querybuilder._makequery(
+            query=query,
+            custom_query={'Title': {'query': 'Test Folder'},
+                          'portal_type': {'query': 'Folder'}})
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].Title(), 'Test Folder')
+
+
+class TestConfigurationFetcher(unittest.TestCase):
+
+    layer = TEST_PROFILE_PLONEAPPQUERYSTRING_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        self.portal.invokeFactory("Folder",
+                                  "testfolder",
+                                  title="Test Folder")
+        self.folder = self.portal.testfolder
 
     def testGettingJSONConfiguration(self):
         self.folder.restrictedTraverse('@@querybuilderjsonconfig')()