+++ plone/app/querystring/tests/testQueryParser.py
@@ -1,14 +1,18 @@
-import unittest
-
 from DateTime import DateTime
 from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.querystring import queryparser
 from plone.registry import field
 from plone.registry import Record
+from plone.registry import Registry
+from plone.registry.interfaces import IRegistry
 
-from .base import UnittestWithRegistryLayer
 from plone.app.querystring.queryparser import Row
+from plone.app.querystring.testing import NOT_INSTALLED_PLONEAPPQUERYSTRING_INTEGRATION_TESTING
+
+from zope.component import getGlobalSiteManager
 from zope.interface.declarations import implements
+
+import unittest2 as unittest
 
 MOCK_SITE_ID = "site"
 
@@ -17,13 +21,13 @@ class MockObject(object):
 
     def __init__(self, uid, path):
         self.uid = uid
-        self.path = path.split("/")
+        self.path = path
 
     def getPath(self):
         return self.path
 
     def getPhysicalPath(self):
-        return self.path
+        return self.path.split("/")
 
     def absolute_url(self):
         return self.path
@@ -84,13 +88,17 @@ class MockSite(object):
 
 class MockUser(object):
 
-    def __init__(self, username=None):
+    def __init__(self, username=None, roles=None):
         self.username = 'Anonymous User'
         if username:
             self.username = username
+        self.roles = roles or "Anonymous"
 
     def getUserName(self):
         return self.username
+
+    def getRoles(self):
+        return self.roles
 
 
 class MockPortal_membership(object):
@@ -104,10 +112,12 @@ class MockPortal_membership(object):
 
 class TestQueryParserBase(unittest.TestCase):
 
-    layer = UnittestWithRegistryLayer
+    layer = NOT_INSTALLED_PLONEAPPQUERYSTRING_INTEGRATION_TESTING
 
     def setUp(self):
-        super(TestQueryParserBase, self).setUp()
+        gsm = getGlobalSiteManager()
+        self.registry = Registry()
+        gsm.registerUtility(self.registry, IRegistry)
         self.setFunctionForOperation(
             'plone.app.querystring.operation.string.is.operation',
             'plone.app.querystring.queryparser._equal')
@@ -119,7 +129,7 @@ class TestQueryParserBase(unittest.TestCase):
         function_field = field.ASCIILine(title=u"Operator")
         function_record = Record(function_field)
         function_record.value = function
-        self.layer.registry.records[operation] = function_record
+        self.registry.records[operation] = function_record
 
 
 class TestQueryParser(TestQueryParserBase):
@@ -140,7 +150,8 @@ class TestQueryParser(TestQueryParserBase):
             'v': '/foo',
         }
         parsed = queryparser.parseFormquery(MockSite(), [data, ])
-        self.assertEqual(parsed, {'path': {'query': '/%s/foo' % MOCK_SITE_ID}})
+        self.assertEqual(
+            parsed, {'path': {'query': ['/%s/foo' % MOCK_SITE_ID]}})
 
     def test_path_computed(self):
         data = {
@@ -148,45 +159,162 @@ class TestQueryParser(TestQueryParserBase):
             'o': 'plone.app.querystring.operation.string.path',
             'v': '00000000000000001',
         }
+        parsed = queryparser.parseFormquery(MockSite(), [data, ])
+        self.assertEqual(
+            parsed, {'path': {'query': ['/%s/foo' % MOCK_SITE_ID]}})
+
+    def test_path_with_depth_computed(self):
+        data = {
+            'i': 'path',
+            'o': 'plone.app.querystring.operation.string.path',
+            'v': '/foo::2',
+        }
 
         parsed = queryparser.parseFormquery(MockSite(), [data, ])
-        self.assertEqual(parsed, {'path': {'query': '/%s/foo' % MOCK_SITE_ID}})
+        self.assertEqual(parsed, {
+            'path': {
+                'query': ['/%s/foo' % MOCK_SITE_ID],
+                'depth': 2
+            }
+        })
+
+    def test_multi_path(self):
+        data_1 = {
+            'i': 'path',
+            'o': 'plone.app.querystring.operation.string.path',
+            'v': '/foo',
+        }
+        data_2 = {
+            'i': 'path',
+            'o': 'plone.app.querystring.operation.string.path',
+            'v': '/bar',
+        }
+
+        parsed = queryparser.parseFormquery(MockSite(), [data_1, data_2])
+        self.assertEqual(
+            parsed, {'path': {'query': [
+                '/%s/foo' % MOCK_SITE_ID,
+                '/%s/bar' % MOCK_SITE_ID]}})
+
+    def test_multi_path_with_depth_computet(self):
+        data_1 = {
+            'i': 'path',
+            'o': 'plone.app.querystring.operation.string.path',
+            'v': '/foo::2',
+        }
+        data_2 = {
+            'i': 'path',
+            'o': 'plone.app.querystring.operation.string.path',
+            'v': '/bar::5',
+        }
+
+        parsed = queryparser.parseFormquery(MockSite(), [data_1, data_2])
+        self.assertEqual(
+            parsed, {'path': {'query': [
+                '/%s/foo' % MOCK_SITE_ID,
+                '/%s/bar' % MOCK_SITE_ID], 'depth': 2}})
 
 
 class TestQueryGenerators(TestQueryParserBase):
 
     def test__between(self):
-        data = Row(index='modified',
-                  operator='_between',
-                  values=['2009/08/12', '2009/08/14'])
+        data = Row(
+            index='modified',
+            operator='_between',
+            values=['2009/08/12', '2009/08/14']
+        )
         parsed = queryparser._between(MockSite(), data)
         expected = {'modified': {'query': ['2009/08/12', '2009/08/14'],
                     'range': 'minmax'}}
         self.assertEqual(parsed, expected)
 
     def test__between_reversed_dates(self):
-        data = Row(index='modified',
-                  operator='_between',
-                  values=['2009/08/14', '2009/08/12'])
+        data = Row(
+            index='modified',
+            operator='_between',
+            values=['2009/08/14', '2009/08/12']
+        )
         parsed = queryparser._between(MockSite(), data)
         expected = {'modified': {'query': ['2009/08/12', '2009/08/14'],
                     'range': 'minmax'}}
         self.assertEqual(parsed, expected)
 
+    def test__equal(self):
+        data = Row(
+            index='modified',
+            operator='_equal',
+            values='2010/03/18'
+        )
+        parsed = queryparser._equal(MockSite(), data)
+        expected = {'modified': {'query': '2010/03/18'}}
+        self.assertEqual(parsed, expected)
+
+        data = Row(
+            index='modified',
+            operator='_equal',
+            values=['2010/03/18', '2010/03/19']
+        )
+        parsed = queryparser._equal(MockSite(), data)
+        expected = {'modified': {'query': ['2010/03/18', '2010/03/19']}}
+        self.assertEqual(parsed, expected)
+
+    def test__intEqual(self):
+        data = Row(
+            index='modified',
+            operator='_intEqual',
+            values='20'
+        )
+        parsed = queryparser._intEqual(MockSite(), data)
+        expected = {'modified': {'query': 20}}
+        self.assertEqual(parsed, expected)
+
+        data = Row(
+            index='modified',
+            operator='_intEqual',
+            values=['20', '21']
+        )
+        parsed = queryparser._intEqual(MockSite(), data)
+        expected = {'modified': {'query': [20, 21]}}
+        self.assertEqual(parsed, expected)
+
+    def test__lessThan(self):
+        data = Row(
+            index='modified',
+            operator='_lessThan',
+            values='2010/03/18'
+        )
+        parsed = queryparser._lessThan(MockSite(), data)
+        expected = {'modified': {'query': '2010/03/18', 'range': 'max'}}
+        self.assertEqual(parsed, expected)
+
+    def test__intLessThan(self):
+        data = Row(
+            index='modified',
+            operator='_intLessThan',
+            values='20'
+        )
+        parsed = queryparser._intLessThan(MockSite(), data)
+        expected = {'modified': {'query': 20, 'range': 'max'}}
+        self.assertEqual(parsed, expected)
+
     def test__largerThan(self):
-        data = Row(index='modified',
-                  operator='_largerThan',
-                  values='2010/03/18')
+        data = Row(
+            index='modified',
+            operator='_largerThan',
+            values='2010/03/18'
+        )
         parsed = queryparser._largerThan(MockSite(), data)
         expected = {'modified': {'query': '2010/03/18', 'range': 'min'}}
         self.assertEqual(parsed, expected)
 
-    def test__lessThan(self):
-        data = Row(index='modified',
-                  operator='_lessThan',
-                  values='2010/03/18')
-        parsed = queryparser._lessThan(MockSite(), data)
-        expected = {'modified': {'query': '2010/03/18', 'range': 'max'}}
+    def test__intLargerThan(self):
+        data = Row(
+            index='modified',
+            operator='_intLargerThan',
+            values='20'
+        )
+        parsed = queryparser._intLargerThan(MockSite(), data)
+        expected = {'modified': {'query': 20, 'range': 'min'}}
         self.assertEqual(parsed, expected)
 
     def test__currentUser(self):
@@ -194,9 +322,11 @@ class TestQueryGenerators(TestQueryParserBase):
         u = MockUser()
         pm = MockPortal_membership(user=u)
         context = MockSite(portal_membership=pm)
-        data = Row(index='Creator',
-                  operator='_currentUser',
-                  values=None)
+        data = Row(
+            index='Creator',
+            operator='_currentUser',
+            values=None
+        )
         parsed = queryparser._currentUser(context, data)
         expected = {'Creator': {'query': 'Anonymous User'}}
         self.assertEqual(parsed, expected)
@@ -205,11 +335,38 @@ class TestQueryGenerators(TestQueryParserBase):
         u = MockUser(username='admin')
         pm = MockPortal_membership(user=u)
         context = MockSite(portal_membership=pm)
-        data = Row(index='Creator',
-                  operator='_currentUser',
-                  values=None)
+        data = Row(
+            index='Creator',
+            operator='_currentUser',
+            values=None
+        )
         parsed = queryparser._currentUser(context, data)
         expected = {'Creator': {'query': 'admin'}}
+        self.assertEqual(parsed, expected)
+
+    def test__showInactive(self):
+        # Anonymous user
+        u = MockUser()
+        pm = MockPortal_membership(user=u)
+        context = MockSite(portal_membership=pm)
+        data = Row(index='show_inactive',
+                   operator='_showInactive',
+                   values=["Manager"])
+        parsed = queryparser._showInactive(context, data)
+        # False is expected since Anonymous doesn't have Manager role
+        expected = {'show_inactive': False}
+        self.assertEqual(parsed, expected)
+
+        # Logged in user 'admin'
+        u = MockUser(username='admin', roles=("Manager",))
+        pm = MockPortal_membership(user=u)
+        context = MockSite(portal_membership=pm)
+        data = Row(index='show_inactive',
+                   operator='_showInactive',
+                   values=["Manager"])
+        parsed = queryparser._showInactive(context, data)
+        # True is expected since Admin should have Manager role
+        expected = {'show_inactive': True}
         self.assertEqual(parsed, expected)
 
     def test__lessThanRelativeDate(self):
@@ -218,9 +375,11 @@ class TestQueryGenerators(TestQueryParserBase):
         mydate = now + days
         expected_dates = [now.earliestTime(), mydate.latestTime()]
         expected = {'modified': {'query': expected_dates, 'range': 'minmax'}}
-        data = Row(index='modified',
-                  operator='_lessThanRelativeDate',
-                  values=days)
+        data = Row(
+            index='modified',
+            operator='_lessThanRelativeDate',
+            values=days
+        )
         parsed = queryparser._lessThanRelativeDate(MockSite(), data)
         self.assertEqual(parsed, expected)
 
@@ -230,9 +389,11 @@ class TestQueryGenerators(TestQueryParserBase):
         mydate = now - days
         expected_dates = [mydate.earliestTime(), now.latestTime()]
         expected = {'modified': {'query': expected_dates, 'range': 'minmax'}}
-        data = Row(index='modified',
-                  operator='_moreThanRelativeDate',
-                  values=days)
+        data = Row(
+            index='modified',
+            operator='_moreThanRelativeDate',
+            values=days
+        )
         parsed = queryparser._moreThanRelativeDate(MockSite(), data)
         self.assertEqual(parsed, expected)
 
@@ -240,27 +401,33 @@ class TestQueryGenerators(TestQueryParserBase):
         now = DateTime()
         expected_dates = [now.earliestTime(), now.latestTime()]
         expected = {'modified': {'query': expected_dates, 'range': 'minmax'}}
-        data = Row(index='modified',
-                  operator='_today',
-                  values=expected_dates)
+        data = Row(
+            index='modified',
+            operator='_today',
+            values=expected_dates
+        )
         parsed = queryparser._today(MockSite(), data)
         self.assertEqual(parsed, expected)
 
     def test__path(self):
         # normal path
-        data = Row(index='path',
-                  operator='_path',
-                  values='/news/')
+        data = Row(
+            index='path',
+            operator='_path',
+            values='/news/'
+        )
         parsed = queryparser._path(MockSite(), data)
-        expected = {'path': {'query': '/%s/news/' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s/news/' % MOCK_SITE_ID]}}
         self.assertEqual(parsed, expected)
 
         # by uid
-        data = Row(index='path',
-                  operator='_path',
-                  values='00000000000000001')
+        data = Row(
+            index='path',
+            operator='_path',
+            values='00000000000000001'
+        )
         parsed = queryparser._path(MockSite(), data)
-        expected = {'path': {'query': '/%s/foo' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s/foo' % MOCK_SITE_ID]}}
         self.assertEqual(parsed, expected)
 
     def test__relativePath(self):
@@ -283,27 +450,31 @@ class TestQueryGenerators(TestQueryParserBase):
                                             path="/%s/bar/egg" % MOCK_SITE_ID)
 
         # show my siblings
-        data = Row(index='path',
-                  operator='_relativePath',
-                  values='..')
+        data = Row(
+            index='path',
+            operator='_relativePath',
+            values='..'
+        )
         parsed = queryparser._relativePath(context, data)
-        expected = {'path': {'query': '/%s/bar' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s/bar' % MOCK_SITE_ID]}}
         self.assertEqual(parsed, expected)
 
         # walk upwards
-        data = Row(index='path',
-                  operator='_relativePath',
-                  values='../../')
+        data = Row(
+            index='path',
+            operator='_relativePath',
+            values='../../'
+        )
         parsed = queryparser._relativePath(context, data)
-        expected = {'path': {'query': '/%s' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s' % MOCK_SITE_ID]}}
         self.assertEqual(parsed, expected)
 
         # if you walk beyond INavigatinRoot it should stop and return
         data = Row(index='path',
-                  operator='_relativePath',
-                  values='../../../')
+                   operator='_relativePath',
+                   values='../../../')
         parsed = queryparser._relativePath(context, data)
-        expected = {'path': {'query': '/%s' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s' % MOCK_SITE_ID]}}
         self.assertEqual(parsed, expected)
 
         # reach a subfolder on Plone root
@@ -311,7 +482,7 @@ class TestQueryGenerators(TestQueryParserBase):
                    operator='_relativePath',
                    values='../../ham')
         parsed = queryparser._relativePath(context, data)
-        expected = {'path': {'query': '/%s/ham' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s/ham' % MOCK_SITE_ID]}}
         self.assertEqual(parsed, expected)
 
         # reach a subfolder on parent of collection
@@ -319,9 +490,19 @@ class TestQueryGenerators(TestQueryParserBase):
                    operator='_relativePath',
                    values='../egg')
         parsed = queryparser._relativePath(context, data)
-        expected = {'path': {'query': '/%s/bar/egg' % MOCK_SITE_ID}}
+        expected = {'path': {'query': ['/%s/bar/egg' % MOCK_SITE_ID]}}
+        self.assertEqual(parsed, expected)
+
+        # relative path with depth
+        data = Row(
+            index='path',
+            operator='_relativePath',
+            values='..::2'
+        )
+        parsed = queryparser._relativePath(context, data)
+        expected = {'path': {'query': ['/%s/bar' % MOCK_SITE_ID], 'depth': 2}}
         self.assertEqual(parsed, expected)
 
     def test_getPathByUID(self):
         actual = queryparser.getPathByUID(MockSite(), '00000000000000001')
-        self.assertEqual(actual, ['', 'site', 'foo'])
+        self.assertEqual(actual, '/site/foo')